# src/main/java/org/brosers/exchanger/openapi/src/main/resources/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: bartosz.olcha1@pl.ey.com
    name: Bartosz Olcha
  description: |-
    EY Mediator BFF REST APIs for mediation and negotiation workflow process handling
    Release notes: - Dictionaries added (REST and structure)
  title: EY Mediator
  version: 0.1.7.17
servers:
- url: /
security:
- bearerAuth: []
paths:
...
  /register/import:
    description: The REST endpoint/path used to send excel file to import caseData.
    post:
      deprecated: true
      description: Send excel file with caseData to import
      operationId: importRegister
      requestBody:
        content:
          application/octet-stream:
            schema:
              format: binary
              type: string
        description: A caseDate register to be imported
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  type: integer
                type: array
          description: Successful response.
      summary: Send excel file to import
      x-content-type: application/octet-stream
      x-accepts: application/json
    summary: Path used to send excel file with register to import
  /register/import-new:
    description: The REST endpoint/path used to send excel file to import caseData.
    post:
      description: Send excel file with caseData to import
      operationId: importRegisterNew
      requestBody:
        content:
          application/octet-stream:
            schema:
              format: binary
              type: string
        description: A caseDate register to be imported
        required: true
      responses:
        "200":
          description: Successful response
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportErrorResponse'
          description: Error in import occurred
      summary: Send excel file to import
      x-content-type: application/octet-stream
      x-accepts: application/json
    summary: Path used to send excel file with register to import
...
components:
  schemas:
    ImportErrorResponse:
      description: Excel import error with list of problems
      example:
        errorList:
        - line: 0
          message: message
        - line: 0
          message: message
        message: message
      properties:
        message:
          type: string
        errorList:
          items:
            $ref: '#/components/schemas/RowProcessingError'
          type: array
      type: object
    RowProcessingError:
      description: Import error description
      example:
        line: 0
        message: message
      properties:
        line:
          format: int64
          type: integer
        message:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http